Sprint 5: Show logging output from tests, new empty file templates, adopt Orion 4.0GA, bugfixes.

+++ bundles/org.eclipse.orion.client.git/web/orion/git/gitRepositoryExplorer.js
@@ -9,11 +9,17 @@
  * Contributors: IBM Corporation - initial API and implementation
  ******************************************************************************/
 /*global define document Image window*/
-define(['i18n!git/nls/gitmessages', 'require', 'orion/commandRegistry', 'orion/section', 'orion/i18nUtil', 'orion/webui/littlelib', 'orion/PageUtil', 'orion/dynamicContent', 'orion/fileUtils', 
+define(['i18n!git/nls/gitmessages', 'require', 'orion/commandRegistry', 'orion/section', 'orion/i18nUtil', 'orion/webui/littlelib', 'orion/git/util', 'orion/URITemplate', 'orion/PageUtil', 'orion/dynamicContent', 'orion/fileUtils', 
         'orion/globalCommands', 'orion/git/gitCommands', 'orion/Deferred', 'orion/git/widgets/CommitTooltipDialog'], 
-		function(messages, require, mCommands, mSection, i18nUtil, lib, PageUtil, mDynamicContent, mFileUtils, mGlobalCommands, mGitCommands, Deferred,
+		function(messages, require, mCommands, mSection, i18nUtil, lib, mGitUtil, URITemplate, PageUtil, mDynamicContent, mFileUtils, mGlobalCommands, mGitCommands, Deferred,
 				mCommitTooltip) {
 var exports = {};
+	
+var repoTemplate = new URITemplate("git/git-repository.html#{,resource,params*}"); //$NON-NLS-0$
+var repoPageTemplate = new URITemplate("git/git-repository.html#{,resource,params*}?page=1&pageSize=20"); //$NON-NLS-0$
+var statusTemplate = new URITemplate(mGitUtil.statusUILocation + "#{,resource,params*}"); //$NON-NLS-0$
+var logTemplate = new URITemplate("git/git-log.html#{,resource,params*}?page=1"); //$NON-NLS-0$
+var commitTemplate = new URITemplate("git/git-commit.html#{,resource,params*}?page=1&pageSize=1"); //$NON-NLS-0$
 
 exports.GitRepositoryExplorer = (function() {
 	
@@ -128,6 +134,9 @@ exports.GitRepositoryExplorer = (function() {
 					lib.empty(pageNav);
 					that.commandService.renderCommands(that.pageNavId, pageNav, resp, that, "button");
 				}
+				if (!resp.Children) {
+					return;
+				}
 				
 				if (resp.Children.length === 0) {
 					that.initTitleBar({});
@@ -241,7 +250,7 @@ exports.GitRepositoryExplorer = (function() {
 		updatePageActions(that.registry, that.commandService, "pageActions", scopeId, repository || {}); //$NON-NLS-1$ //$NON-NLS-0$
 		mGlobalCommands.setPageTarget({task: messages.Repo, target: repository, breadcrumbTarget: item,
 			makeBreadcrumbLink: function(seg, location) {
-				seg.href = require.toUrl("git/git-repository.html") + (location ? "#" + location : ""); //$NON-NLS-0$
+				seg.href = require.toUrl(repoTemplate.expand({resource: location || ""}));
 			},
 			serviceRegistry: this.registry, commandService: this.commandService}); 
 	};
@@ -368,6 +377,9 @@ exports.GitRepositoryExplorer = (function() {
 		
 			initialRender : function(){
 				var tableNode = lib.node('table');	 //$NON-NLS-0$
+				if (!tableNode) {
+					return;
+				}
 				lib.empty(tableNode);
 				
 				if(!repositories || repositories.length === 0){
@@ -410,7 +422,7 @@ exports.GitRepositoryExplorer = (function() {
 
 				if (links){
 					var link = document.createElement("a");
-					link.href = require.toUrl("git/git-repository.html#") + repositories[i].Location;
+					link.href = require.toUrl(repoTemplate.expand({resource: repositories[i].Location}));
 					link.appendChild(document.createTextNode(repositories[i].Name));
 					title.appendChild(link);
 				} else { 
@@ -473,7 +485,11 @@ exports.GitRepositoryExplorer = (function() {
 	};
 	
 	GitRepositoryExplorer.prototype.renderRepository = function(repository, index, length, mode, links){
-		lib.node("location"+index).textContent = messages["location: "] + repository.Content.Path;
+		var locationnode = lib.node("location"+index);
+		if (!locationnode) {
+			return;
+		}
+		locationnode.textContent = messages["location: "] + repository.Content.Path;
 		var status = repository.Status;
 		
 		if (mode === "full"){ //$NON-NLS-0$
@@ -514,7 +530,7 @@ exports.GitRepositoryExplorer = (function() {
 		
 		that.commandService.registerCommandContribution(titleWrapper.actionsNode.id, "eclipse.orion.git.repositories.viewAllCommand", 10); //$NON-NLS-0$
 		that.commandService.renderCommands(titleWrapper.actionsNode.id, titleWrapper.actionsNode.id, 
-			{"ViewAllLink":"git/git-status.html#" + repository.StatusLocation, "ViewAllLabel": messages["See Full Status"], "ViewAllTooltip": messages["See the status"]}, that, "button");
+			{"ViewAllLink":statusTemplate.expand({resource: repository.StatusLocation}), "ViewAllLabel": messages["See Full Status"], "ViewAllTooltip": messages["See the status"]}, that, "button"); //$NON-NLS-3$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
 		
 		var progress = titleWrapper.createProgressMonitor();
 		progress.begin("Loading status"); //$NON-NLS-0$
@@ -618,7 +634,7 @@ exports.GitRepositoryExplorer = (function() {
 				if (mode !== "full" && branches.length !== 0){ //$NON-NLS-0$
 					that.commandService.registerCommandContribution(titleWrapper.actionsNode.id, "eclipse.orion.git.repositories.viewAllCommand", 10); //$NON-NLS-0$
 					that.commandService.renderCommands(titleWrapper.actionsNode.id, titleWrapper.actionsNode.id, 
-						{"ViewAllLink":"git/git-repository.html#" + branchLocation + "?page=1", "ViewAllLabel":messages['View All'], "ViewAllTooltip":messages["View all local and remote tracking branches"]}, that, "button"); //$NON-NLS-7$ //$NON-NLS-5$ //$NON-NLS-3$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
+						{"ViewAllLink":repoTemplate.expand({resource: branchLocation}), "ViewAllLabel":messages['View All'], "ViewAllTooltip":messages["View all local and remote tracking branches"]}, that, "button"); //$NON-NLS-3$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
 				}
 				
 				that.commandService.registerCommandContribution(titleWrapper.actionsNode.id, "eclipse.addBranch", 200); //$NON-NLS-0$
@@ -652,7 +668,7 @@ exports.GitRepositoryExplorer = (function() {
 
 		if (branch.Current){
 			var span = document.createElement("span");
-			span.className = "sectionIcon gitImageSprite git-sprite-branch_active";
+			span.className = "sectionIcon gitImageSprite git-sprite-branch-active";
 			horizontalBox.appendChild(span);
 		}
 		
@@ -826,7 +842,7 @@ exports.GitRepositoryExplorer = (function() {
 
 				that.commandService.registerCommandContribution(titleWrapper.actionsNode.id, "eclipse.orion.git.repositories.viewAllCommand", 10); //$NON-NLS-0$
 				that.commandService.renderCommands(titleWrapper.actionsNode.id, titleWrapper.actionsNode.id, 
-					{"ViewAllLink":"git/git-log.html#" + currentBranch.CommitLocation + "?page=1", "ViewAllLabel":messages["See Full Log"], "ViewAllTooltip":messages["See the full log"]}, that, "button"); //$NON-NLS-7$ //$NON-NLS-6$ //$NON-NLS-5$ //$NON-NLS-3$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
+					{"ViewAllLink":logTemplate.expand({resource: currentBranch.CommitLocation}), "ViewAllLabel":messages["See Full Log"], "ViewAllTooltip":messages["See the full log"]}, that, "button"); //$NON-NLS-7$ //$NON-NLS-6$ //$NON-NLS-5$ //$NON-NLS-3$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
 				
 				if (tracksRemoteBranch){
 					that.commandService.registerCommandContribution(titleWrapper.actionsNode.id, "eclipse.orion.git.fetch", 100); //$NON-NLS-0$
@@ -939,7 +955,7 @@ exports.GitRepositoryExplorer = (function() {
 		horizontalBox.style.overflow = "hidden";
 		sectionItem.appendChild(horizontalBox);
 
-		var imgSpriteName = (outgoing ? "git-sprite-outgoing_commit" : "git-sprite-incoming_commit"); //$NON-NLS-1$ //$NON-NLS-0$
+		var imgSpriteName = (outgoing ? "git-sprite-outgoing-commit" : "git-sprite-incoming-commit"); //$NON-NLS-1$ //$NON-NLS-0$
 		
 		var span = document.createElement("span");
 		span.className = "sectionIcon gitImageSprite " + imgSpriteName;
@@ -962,7 +978,7 @@ exports.GitRepositoryExplorer = (function() {
 
 		var titleLink = document.createElement("a");
 		titleLink.className = "navlinkonpage";
-		titleLink.href = require.toUrl("git/git-commit.html#") + commit.Location + "?page=1&pageSize=1";
+		titleLink.href = require.toUrl(commitTemplate.expand({resource: commit.Location}));
 		titleLink.textContent = commit.Message;
 		detailsView.appendChild(titleLink);
 
@@ -1041,16 +1057,19 @@ exports.GitRepositoryExplorer = (function() {
 				var tagRenderer = {
 					initialRender : function(){
 						progress.done();
-						
-						lib.empty(lib.node("tagNode"));
-						lib.node("tagNode").appendChild(tagsContainer);
+						var tagNode = lib.node("tagNode");
+						if (!tagNode) {
+							return;
+						}
+						lib.empty(tagNode);
+						tagNode.appendChild(tagsContainer);
 						
 						that.commandService.destroy(titleWrapper.actionsNode.id);
 						
 						if (mode !== "full" && tags.length !== 0){ //$NON-NLS-0$
 							that.commandService.registerCommandContribution(titleWrapper.actionsNode.id, "eclipse.orion.git.repositories.viewAllCommand", 10); //$NON-NLS-0$
 							that.commandService.renderCommands(titleWrapper.actionsNode.id, titleWrapper.actionsNode.id,
-									{"ViewAllLink":require.toUrl("git/git-repository.html#") + repository.TagLocation + "?page=1&pageSize=20", "ViewAllLabel":messages['View All'], "ViewAllTooltip":messages["View all tags"]}, that, "button"); //$NON-NLS-7$ //$NON-NLS-5$ //$NON-NLS-3$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
+									{"ViewAllLink":require.toUrl(repoPageTemplate.expand({resource: repository.TagLocation})), "ViewAllLabel":messages['View All'], "ViewAllTooltip":messages["View all tags"]}, that, "button"); //$NON-NLS-7$ //$NON-NLS-5$ //$NON-NLS-3$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
 						}
 		
 						if (tags.length === 0) {
@@ -1115,7 +1134,7 @@ exports.GitRepositoryExplorer = (function() {
 		
 		var link = document.createElement("a");
 		link.className = "navlinkonpage";
-		link.href = require.toUrl("git/git-commit.html#") + commit.Location + "?page=1&pageSize=1";
+		link.href = require.toUrl(commitTemplate.expand({resource: commit.Location}));
 		link.textContent = commit.Message;
 		description.appendChild(link);
 		
@@ -1234,7 +1253,7 @@ exports.GitRepositoryExplorer = (function() {
 
 					that.commandService.registerCommandContribution(titleWrapper.actionsNode.id, "eclipse.orion.git.repositories.viewAllCommand", 10); //$NON-NLS-0$
 					that.commandService.renderCommands(titleWrapper.actionsNode.id, titleWrapper.actionsNode.id,
-							{"ViewAllLink":"git/git-repository.html#" + configLocation, "ViewAllLabel":messages['View All'], "ViewAllTooltip":messages["View all configuration entries"]}, that, "button"); //$NON-NLS-6$ //$NON-NLS-4$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
+							{"ViewAllLink":repoTemplate.expand({resource: configLocation}), "ViewAllLabel":messages['View All'], "ViewAllTooltip":messages["View all configuration entries"]}, that, "button"); //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-2$ //$NON-NLS-1$ //$NON-NLS-0$
 				}
 				
 				if (mode === "full"){ //$NON-NLS-0$
